Refactor postDetail
