+++ src/Components/Posts/PostDetail/Renderers/index.js
@@ -1,0 +1,26 @@
+import React from "react";
+
+const flatten = (text, child) => {
+  return typeof child === "string"
+    ? text + child
+    : React.Children.toArray(child.props.children).reduce(flatten, text);
+};
+
+export const HeadingRenderer = props => {
+  var children = React.Children.toArray(props.children);
+  var text = children.reduce(flatten, "");
+  var slug = text.toLowerCase().replace(/[!?\s]/g, "-");
+  return React.createElement(
+    "h" + props.level,
+    { id: slug, class: "anchor" },
+    props.children
+  );
+};
+
+export const formatDate = date => {
+  const dateObject = new Date(date);
+  const month = dateObject.getMonth() + 1;
+  const day = dateObject.getDate();
+  const year = dateObject.getFullYear();
+  return `${year}/${month}/${day}`;
+};
